Implement Ceilometer and AODH controls based on Cinder
The OpenStack Security guide does not currently include
a checklist of controls for Ceilometer and AODH, so the controls were added
based on what glance and cinder do. Since Ceilometer or AODH
is usually optional, the controls will not run unless
ceilometer_enabled or aodh_enabled is specified to
be true by attributes.